LUCENESOLR-1811: Remove client-close logic in SRI
StreamingResultsIterator had code that conditionally closed
HttpSolrClient's in its hasNext() method.  This code originally was
copied from a different project, where this decision made a little more
sense. But in zeppelin-solr, SolrClients are always tied to the
lifecycle of the interpreter and should never be closed prematurely.

This commit removes the errant closure, which was causing any request
after a zero-result query to error out.